修改Character座標

+++ game4.10/Source/Map_Home.cpp
@@ -97,7 +97,7 @@ namespace game_framework {
 		vector<Enemy*>::iterator iter;
 		for (iter = _enemies.begin(); iter != _enemies.end(); iter++)
 		{
-			int *e_xy = (*iter)->GetPosition();
+			float *e_xy = (*iter)->GetPosition();
 			int *e_collision_move = (*iter)->GetCollisionMove();
 
 			int x1 = _cxy[0] + collision_move[0] + dx;
